new module refactoring
