+++ app/lib/helpers.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# Candidates controller class
+# Helpers methods
 module Helpers
   # Authentification
   def current_user
@@ -75,14 +75,6 @@ module Helpers
     end
   end
 
-  # Form-s helpers
-  def intitalise_form_variables
-    @last_job_like_dislike_params = last_job_like_dislike_params
-    @work_experience_areas        = work_experience_areas
-    @desired_pay_system           = desired_pay_system
-    @tables_names                 = tables_names
-  end
-
   def tables_names
     {
       'relatives' => :name,
@@ -91,43 +83,6 @@ module Helpers
       'language' => :name,
       'experience' => :name,
       'reccomenders' => :name
-    }
-  end
-
-  # Abbreviations
-  def last_job_like_dislike_params
-    {
-      'ls' => 'Низкая зарплата',
-      'upct' => 'Неудовлетворительный психологический климат в коллективе',
-      'llbo' => 'Невысокий уровень организации дела',
-      'drvm' => 'Сложные отношения с руководством',
-      'ncp' => 'Нет перспективы должностного роста',
-      'emr' => 'Чрезмерно высокие требования руководства',
-      'ow' => 'Сверхурочная работа',
-      'so' => 'Что-то другое'
-    }
-  end
-
-  def work_experience_areas
-    {
-      'prod' => 'Производство',
-      'serv' => 'Услуги',
-      'whsal' => 'Оптовая торговля',
-      'ret' => 'Розничная торговля',
-      'publ' => 'Издательство',
-      'pc' => 'Общепит',
-      'build' => 'Строительство',
-      'tr' => 'Транспорт',
-      'ent' => 'Индивидуальный предприниматель'
-    }
-  end
-
-  def desired_pay_system
-    {
-      sal: 'Оклад',
-      salbon: 'Оклад+премия',
-      int: 'Процент',
-      salint: 'Оклад+процент'
     }
   end
 end
